+++ main/bundle_of_joy/lib/appointmentMother/appointmentMother_main.dart
@@ -1,7 +1,7 @@
-import "package:flutter/material.dart";
-import "appointmentMother_add_1.dart";
-import "appointmentMother_add_full.dart";
-import "appointmentMother_recordList.dart";
+import 'package:bundle_of_joy/widgets/horizontalCardWidget.dart';
+import 'package:flutter/material.dart';
+import 'appointmentMother_add_1_hospital.dart';
+import 'appointmentMother_recordList.dart';
 
 class AppointmentMotherMain extends StatefulWidget {
   @override
@@ -9,126 +9,70 @@ class AppointmentMotherMain extends StatefulWidget {
 }
 
 class _AppointmentMotherMainState extends State<AppointmentMotherMain> {
-  // BUILD THE WIDGET
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      // APP BAR
-      appBar: AppBar(
-        toolbarHeight: MediaQuery.of(context).size.height * 0.1,
-        title: Text(
-          "Appointment Management",
-          style: TextStyle(
-            fontFamily: 'Comfortaa',
-            fontWeight: FontWeight.bold,
-            fontSize: MediaQuery.of(context).size.width * 0.05,
-            color: Colors.black,
+      backgroundColor: Color(0xFFf5f5f5),
+      body: CustomScrollView(
+        physics: const BouncingScrollPhysics(),
+        slivers: <Widget>[
+          SliverAppBar(
+            expandedHeight: MediaQuery.of(context).size.height * 0.4,
+            floating: true,
+            pinned: true,
+            stretch: true,
+            stretchTriggerOffset: 70.0,
+            flexibleSpace: FlexibleSpaceBar(
+              centerTitle: true,
+              collapseMode: CollapseMode.parallax,
+              stretchModes: [
+                StretchMode.zoomBackground,
+              ],
+              title: Text(
+                "Appointment Management",
+                style: TextStyle(
+                  fontSize: MediaQuery.of(context).size.width * 0.045,
+                ),
+              ),
+              background: Image.network(
+                "https://static.vecteezy.com/system/resources/previews/000/171/284/original/free-hand-drawn-vector-nightscape-illustration.jpg",
+                fit: BoxFit.cover,
+              ),
+            ),
           ),
-        ),
-        iconTheme: IconThemeData(
-          color: Colors.black, //change your color here
-        ),
-        //automaticallyImplyLeading: false, // CENTER THE TEXT
-        backgroundColor: Color(0xFFFCFFD5),
-        centerTitle: true,
-      ),
-
-      // BODY
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            InkWell(
-              child: Container(
-                width: MediaQuery.of(context).size.width * 0.75,
-                height: MediaQuery.of(context).size.height * 0.25,
-                decoration: myBoxDecoration(),
-                child: Container(
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      Container(
-                        margin: EdgeInsets.only(bottom: 20),
-                        child: Image.asset(
-                          "assets/icons/appointment.png",
-                          height: 90,
-                        ),
-                      ),
-                      Container(
-                        child: Text(
-                          "Upcoming Appointments",
-                          style: TextStyle(
-                            fontFamily: 'Comfortaa',
-                            fontWeight: FontWeight.bold,
-                            fontSize: MediaQuery.of(context).size.height * 0.025,
-                            color: Colors.black,
-                          ),
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-              onTap: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) => AppointmentMotherRecordList()),
-                );
-              },
-            ),
-            SizedBox(height: MediaQuery.of(context).size.height * 0.05),
-            Container(
-              width: MediaQuery.of(context).size.width * 0.75,
-              height: MediaQuery.of(context).size.height * 0.25,
-              decoration: myBoxDecoration(),
-              child: InkWell(
-                child: Container(
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      Container(
-                        margin: EdgeInsets.only(bottom: 20),
-                        child: Image.asset(
-                          "assets/icons/add.png",
-                          height: 80,
-                        ),
-                      ),
-                      Container(
-                        child: Text(
-                          "Add Appointments",
-                          style: TextStyle(
-                            fontFamily: 'Comfortaa',
-                            fontWeight: FontWeight.bold,
-                            fontSize: MediaQuery.of(context).size.height * 0.025,
-                            color: Colors.black,
-                          ),
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-                onTap: () {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(builder: (context) => AppointmentMotherAddFull()),
-                  );
-                },
+          SliverFillRemaining(
+            fillOverscroll: true,
+            hasScrollBody: false,
+            child: Padding(
+              padding: const EdgeInsets.all(5),
+              child: Column(
+                children: [
+                  HorizontalCardWidget(
+                      title: "View Appointments",
+                      description: "View and manage all of your appointments.",
+                      svgSrc: "assets/icons/Hamburger.svg",
+                      press: () {
+                        Navigator.push(
+                          context,
+                          MaterialPageRoute(builder: (context) => AppointmentMotherRecordList()),
+                        );
+                      }),
+                  HorizontalCardWidget(
+                      title: "Book An Appoinment",
+                      description: "Book an appointment with the doctor.",
+                      svgSrc: "assets/icons/Hamburger.svg",
+                      press: () {
+                        Navigator.push(
+                          context,
+                          MaterialPageRoute(builder: (context) => AppointmentMotherAddHospital()),
+                        );
+                      }),
+                ],
               ),
             ),
-          ],
-        ),
+          ),
+        ],
       ),
-    );
-  }
-
-  BoxDecoration myBoxDecoration() {
-    return BoxDecoration(
-      color: Color(0xFFFCFFD5),
-      border: Border.all(
-        color: Colors.black,
-        width: 2.0,
-      ),
-      borderRadius: BorderRadius.all(Radius.circular(10.0)),
     );
   }
 }
